Fix warnings when calling FactoryListBox drawItem

+++ OPHD/UI/FactoryListBox.cpp
@@ -163,9 +163,9 @@ void FactoryListBox::update()
 	{
 		drawItem(renderer, *static_cast<FactoryListBoxItem*>(mItems[i]),
 			positionX(),
-			positionY() + (i * LIST_ITEM_HEIGHT),
-			static_cast<float>(item_width()),
-			static_cast<float>(draw_offset()),
+			positionY() + (static_cast<int>(i) * LIST_ITEM_HEIGHT),
+			item_width(),
+			draw_offset(),
 			i == currentSelection());
 	}
